feat: fix sharing option
